refactor: generalize usage
