+++ src/frontend/src/icp/components/info/Info.svelte
@@ -5,13 +5,14 @@
 		tokenCkEthLedger
 	} from '$icp/derived/ic-token.derived';
 	import InfoBitcoin from '$icp/components/info/InfoBitcoin.svelte';
-	import { type HideInfoKey, saveHideInfo, shouldHideInfo } from '$icp/utils/ck.utils';
-	import InfoBox from '$icp/components/info/InfoBox.svelte';
-	import { isNullish, nonNullish } from '@dfinity/utils';
-	import InfoEthereum from '$icp/components/info/InfoEthereum.svelte';
+	import type { HideInfoKey } from '$icp/utils/ck.utils';
+	import InfoBox from '$lib/components/info/InfoBox.svelte';
+	import InfoEthereum from '$lib/components/info/InfoEthereum.svelte';
 	import { isNetworkIdBTCMainnet, isNetworkIdETHMainnet } from '$icp/utils/ic-send.utils';
 	import type { OptionIcCkToken } from '$icp/types/ic';
 	import { token } from '$lib/stores/token.store';
+	import { ckEthereumTwinToken } from '$icp-eth/derived/cketh.derived';
+	import { tokenWithFallback } from '$lib/derived/token.derived';
 
 	let mainnet = true;
 	$: mainnet =
@@ -35,27 +36,14 @@
 			: ckErc20
 				? 'oisy_ic_hide_erc20_info'
 				: undefined;
-
-	let hideInfo = true;
-	$: hideInfo = nonNullish(key) ? shouldHideInfo(key) : true;
-
-	const close = () => {
-		hideInfo = true;
-
-		if (isNullish(key)) {
-			return;
-		}
-
-		saveHideInfo(key);
-	};
 </script>
 
 {#if ckBTC || ckETH || ckErc20}
-	<InfoBox {hideInfo} on:click={close}>
+	<InfoBox {key}>
 		{#if ckBTC}
 			<InfoBitcoin />
 		{:else}
-			<InfoEthereum />
+			<InfoEthereum token={$ckEthereumTwinToken} ckTokenSymbol={$tokenWithFallback.symbol} />
 		{/if}
 	</InfoBox>
 {/if}
